core data階段完成至修改刪除計時暫時可以使用

+++ TransferTravel/Model/RouteModel.swift
@@ -12,7 +12,7 @@ import CoreData
 class RouteItem: NSManagedObject{
 	
 	@NSManaged var routeName : String?
-	@NSManaged var routeID: Int
+	@NSManaged var routeID: Int64
 
 }
